增加新的模块
