+++ readme.txt
@@ -1,0 +1,3 @@
+commonModal：提供公共的基础服务，比如文件上传、邮件发送、权限管理等等
+businessModal：业务模块，是系统真正要实现的业务。依赖于commonModal，比如订单管理、财务统计、会员管理等
+application：可发布的web应用，由各个businessModal组成，最终满足项目的整体需求
