+++ spring-demo-web/src/main/java/com/example/base/controller/BaseController.java
@@ -1,0 +1,5 @@
+package com.example.base.controller;
+
+public abstract class BaseController {
+    //
+}
