+++ spring-demo-web/src/main/java/com/example/profile/Development.java
@@ -1,0 +1,14 @@
+package com.example.profile;
+
+import org.springframework.context.annotation.Profile;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+@Target(ElementType.TYPE)
+@Retention(RetentionPolicy.RUNTIME)
+@Profile("development")
+public @interface Development {
+}
