+++ spring-demo-web/src/main/java/com/example/profile/Test.java
@@ -1,0 +1,14 @@
+package com.example.profile;
+
+import org.springframework.context.annotation.Profile;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+@Profile("test")
+@Retention(RetentionPolicy.RUNTIME)
+@Target(ElementType.TYPE)
+public @interface Test {
+}
