+++ spring-demo-web/src/main/java/com/example/user/PortalIndexController.java
@@ -1,0 +1,13 @@
+package com.example.user;
+
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.RequestMapping;
+
+@Controller
+@RequestMapping("/index")
+public class PortalIndexController {
+    @RequestMapping("")
+    String index(){
+        return "index";
+    }
+}
