+++ spring-demo-web/src/main/java/com/example/user/user/User.java
@@ -1,0 +1,7 @@
+package com.example.user.user;
+
+import com.example.base.entity.Entity;
+
+public class User extends Entity {
+    String loginId;
+}
