+++ spring-demo-web/src/main/java/com/example/user/user/UserServiceInf.java
@@ -1,0 +1,4 @@
+package com.example.user.user;
+
+public interface UserServiceInf {
+}
