+++ spring-demo-web/src/main/webapp/WEB-INF/template/common.html
