--- src/main/java/com/example/annotation/profile/Development.java
@@ -1,14 +1,0 @@
-package com.example.annotation.profile;
-
-import org.springframework.context.annotation.Profile;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-@Target(ElementType.TYPE)
-@Retention(RetentionPolicy.RUNTIME)
-@Profile("development")
-public @interface Development {
-}
