--- src/main/java/com/example/business/user/PortalIndexController.java
@@ -1,13 +1,0 @@
-package com.example.business.user;
-
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.RequestMapping;
-
-@Controller
-@RequestMapping("/index")
-public class PortalIndexController {
-    @RequestMapping("")
-    String index(){
-        return "index";
-    }
-}
