--- src/main/java/com/example/business/user/user/User.java
@@ -1,7 +1,0 @@
-package com.example.business.user.user;
-
-import com.example.mybatisplus.entity.Entity;
-
-public class User extends Entity {
-    String loginId;
-}
