--- src/main/java/com/example/business/user/user/UserServiceInf.java
@@ -1,4 +1,0 @@
-package com.example.business.user.user;
-
-public interface UserServiceInf {
-}
