--- src/main/java/com/example/cache/CacheConfig.java
@@ -1,7 +1,0 @@
-package com.example.cache;
-
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-public class CacheConfig {
-}
