--- src/main/java/com/example/common/base/controller/BaseController.java
@@ -1,5 +1,0 @@
-package com.example.common.base.controller;
-
-public abstract class BaseController {
-    //
-}
