--- src/main/java/com/example/mapper/user/UserMapper.java
@@ -1,6 +1,0 @@
-package com.example.mapper.user;
-
-import com.example.mybatisplus.mapper.Mapper;
-
-public interface  UserMapper extends Mapper {
-}
