--- src/main/java/com/example/shiro/ShiroConfig.java
@@ -1,15 +1,0 @@
-package com.example.shiro;
-
-import com.example.annotation.profile.Development;
-import org.springframework.context.annotation.Configuration;
-
-/**
- * @author lizhi
- * @date 2019/4/8
- * @description shiro配置文件
- */
-@Configuration
-@Development
-public class ShiroConfig {
-
-}
